--- extracted_data/summary.csv
@@ -1,9 +1,0 @@
-index,actual,predictions,delta
-count,503.0,503.0,503.0
-mean,343.55096195895914,343.580248944565,3.877311145335021
-std,70.50049783408235,71.17227328930979,3.2686071678688706
-min,210.61904907226562,210.36459350585938,0.002471923828125
-25%,282.775146484375,282.3796691894531,1.4063262939453125
-50%,336.1766662597656,335.82940673828125,3.090667724609375
-75%,410.47943115234375,411.2623748779297,5.4517974853515625
-max,466.7187805175781,468.0837707519531,19.980682373046875
